+++ src/test/resources/inspection/noInspectionsForValuesInWhereIn.php
@@ -1,0 +1,9 @@
+<?php
+
+class User extends \Illuminate\Database\Eloquent\Model {
+}
+
+User::whereIn('id', [
+    'id1',
+    'id2',
+])->get();